Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    float a,b,c;
+    double a,b,c;
     float res;
-    scanf("%f%f%f",&a,&b, &c);
+    scanf("%lf%lf%lf", &a, &b, &c);
     res = (a+b+c)/3
     printf("Average: %.2f", res);
     return 0;
